Issue #SC-1462: Moved class BaseCacheUpdaterService to core
